refactor: extract bundle instance baking
